Fixed an issue where dependencies in spec choices were not being written to the database correctly casuing the db to throw an error.
Spec maint step categories on the choice level now load with a step category if a step is present in that choice when a spec is loaded into the page.

Removed a bunch of stuff from SPA that is no longer needed like alloy and bake info.  This information is being handled differently on the spec side.

Fixed a couple null reference errors when keeping a SPA after review.

Added more info about choices in SPA maintenence.

Fixed an issue where sample plan entry lines weren't generating properly after 10.