Logging is hooked up to the website and needs to be applied to exceptions w/ HttpContext.TraceIdentifier to track unique errors in the log. (See StepMaintenance.cshtml.cs for example)
Moved logging address to config and added loggin address for staging and production.

Added config to program.cs in website and api to get access to the logging addresses.

Website now has a custom HttpMessageHandler that adds more meaningful data to exceptions for logging.  This class also grants us more control over and visibility of the the requests/responses between website and api.

Added httpContextAccessors to dataAccess so they know what page is requesting what information.  This is helpful for logging but I am not certain if this is the soluition we want to go with.  THIS NEEDS MORE TESTING.

But everything is stable.